+++ tests/profiles/fontwerk_test.py
@@ -11,13 +11,11 @@ def test_check_name_no_mac_entries():
     check = CheckTester(fontwerk_profile,
                         'com.fontwerk/check/no_mac_entries')
 
-    # try fonts with Mac names
     font = TEST_FILE('abeezee/ABeeZee-Italic.ttf')
-    message = assert_results_contain(check(font),
-                           FAIL, 'Mac Names')
+    assert_results_contain(check(font),
+                           FAIL, 'mac-names',
+                           'with a font containing Mac names')
 
-    # try fonts without Mac names
     font = TEST_FILE('source-sans-pro/OTF/SourceSansPro-Regular.otf')
-    assert_PASS(check(font))
-
-
+    assert_PASS(check(font),
+                'with a font without Mac names')