+++ CHANGELOG.md
@@ -2,12 +2,23 @@ Below are the most important changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ## 0.8.6 (2022-Feb-??)
+### New Profile
+  - Olli Meier (@moontypespace) contributed a new profile for Fontwerk, https://fontwerk.com/
+
 ### New Checks
+#### Added to the Fontwerk Profile
+  - **[com.fontwerk/check/no_mac_entries]:** Check if font has Mac name table entries (platform=1) (PR #3545)
+
 #### Added to the Universal Profile
   - **[com.google.fonts/check/designspace_has_sources]:** Check that all sources in a designspace can be loaded successfully. (PR #3168)
   - **[com.google.fonts/check/designspace_has_default_master]:** Check that a default master is defined. (PR #3168)
   - **[com.google.fonts/check/designspace_has_consistent_glyphset]:** Check that non-default masters do not contain glyphs not found in the default master. (PR #3168)
   - **[com.google.fonts/check/designspace_has_consistent_codepoints]:** Check that Unicode assignments are consistent between masters. (PR #3168)
+
+### Changes to existing checks
+#### On the Google Fonts Profile
+  - **[com.google.fonts/check/name/rfn]:** If the OFL text is included in a name table entry, the check should not FAIL, as the full license text contains the term 'Reserved Font Name', which in this case is OK. (issue #3542)
+  - **[com.google.fonts/check/layout_valid_feature_tags]:** Allow 'HARF' and 'BUZZ' tags. (issue #3368)
 
 
 ## 0.8.5 (2022-Jan-13)